requirements.txt
pydantic-ai
ragas
langchain
langchain-ollama>=1.0.6
streamlit>=1.28.0
pandas>=2.0.0
sqlalchemy>=2.0.0
crawl4ai>=0.3.0
loguru>=0.7.0
python-dotenv>=1.0.0
openai>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
playwright>=1.40.0